Explanation of odd squencing of methods in handleLoadMoreResults...
Have seen this in other react apps: if you have dispatch action 1 and
dispatch action 2, and 2 depends on the state change caused by 1,
2 may not recognize the change caused by one.  There is a delay,
particularly if the actions are passed down to children components.

So in handlLoadMoreResults, I inserted a temp variable to stand in
for the increased starting ID, called the fetch request with that
temp variable, and then dispatch the action to increase the
starting ID in the state.  The increased number should then
be available for the next call.

that 1 brought about in the state