New idea for testing on further research: test reducer and action creators together...
... This is not quite unit testing. And some argue you should just test the action
creators that they return the right object and that's it.  But as action creators
receive incoming messages and affect state when dispatched to reducer, it makes
sense to me to do a bit of integration testing here.  Going to try it out in
new test file: reducerAndActionCreators

See here for ideas (some that are different), but author implies, don't worry so
much about testing whether dispatch works, b/c that's redux's job.  That's
what I was concerned about previously, but the above approach seems simpler and
more easily tested.

https://willowtreeapps.com/ideas/best-practices-for-unit-testing-with-a-react-redux-approach