Partial solution to testing getBookRecords with fetchMock...not complete:
Ok, so I've determined a few things here.  Look at both fetchRequestBasicSearch
for re-write of getBookRecords.  And see getBookRecords.test.  I basically had
to find a way to allow any URL to be passed into fetch (with "*") in the test,
and I think no actions were being returned b/c I was providing a response that
was too simple, it wasn't making it through the tests within getBookRecords fetch
request itself.  This proves that mockFetch can be used to see the actions
in getBookRecords though -- I can seem them being passed to getActions()
printed in the console.  To now pass good data response in fetchMock.