Solving Heroku Error - Step 1: Delete Header.js
Here's the error:

Failed to compile.
remote:
remote:        ./src/App.js
remote:        Cannot find file './components_presentational/Header' in './src'
remote:

Here's the problem:

Originally I had named it header and changed it via my text editor, Atom.  Git did not pick up on this filename change.  I only discovered this when I downloaded a branch of the repo and tried to push that to Heroku (I originally though it was a cache problem).

So I have to delete the file, commit, reinsert file with Header.js name, and recommit, and then push to heroku.